[#376] Sonar fixes.

+++ java/lionheart-game/src/main/java/com/b3dgs/lionheart/Hud.java
@@ -67,19 +67,19 @@ public final class Hud implements Resource, Updatable, Renderable
     private static final int HEALTH_Y = 1;
 
     private static final double TALISMENT_X_RATIO = 0.205;
-    private static final int TALISMENT_Y = 1;
+    private static final double TALISMENT_Y = 1.0;
     private static final int TALISMENT_TILE = 0;
 
     private static final double AMULET_X_RATIO = 0.72;
-    private static final int AMULET_Y = 1;
+    private static final double AMULET_Y = 1.0;
     private static final int AMULET_TILE = 1;
 
     private static final double SWORD_X_RATIO = 0.5;
-    private static final int SWORD_Y = 1;
+    private static final double SWORD_Y = 1.0;
     private static final int SWORD_TILE = 2;
 
     private static final int LIFE_TILE = 6;
-    private static final int LIFE_Y = 1;
+    private static final double LIFE_Y = 1.0;
     private static final int LIFE_X_BORDER = 4;
 
     private static final double TIME_X_RATIO = 0.0;
